// src/types.ts
export type SchedulingAlgorithm = 
  | 'FCFS' 
  | 'SJF' 
  | 'SRTF' 
  | 'RR' 
  | 'PRIORITY_NP' 
  | 'PRIORITY_P' 
  | 'MLQ';

export interface AlgorithmDetails {
  id: SchedulingAlgorithm;
  name: string;
  shortName: string;
  description: string;
  group: string;
  requiresTimeQuantum?: boolean;
  requiresPriority?: boolean;
}

export interface Process {
  id: number;
  name: string;
  arrivalTime: number;
  burstTime: number;
  priority?: number;
  completionTime?: number;
  turnaroundTime?: number;
  waitingTime?: number;
}